brv_server/requirements.txt
Flask==2.1.1
Flask-SocketIO==5.1.1
Werkzeug==2.2.2
Litellm==1.25.2
python-docker
docker
pyautogen
pyautogen[redis]
pyautogen[ipython]
pyautogen[graph]
flask_cors
numpy
unstructured
unstructured[all-docs]